break up SpriteFactory into multiple files
Make SpriteFactory a collection of partial classes to make it easier to manage.
