+++ Game/SpriteHandlers/SpriteFactory/ProjectilesSpriteFactory.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Content;
+
+namespace MainGame.SpriteHandlers
+{
+    public static partial class SpriteFactory
+    {
+    }
+}
